requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=0.13.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py36h8f6f2f9_0
ase=3.22.0=pypi_0
astor=0.8.1=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py_1
boost=1.74.0=py36h355b2fd_3
boost-cpp=1.74.0=hc6e9bd1_3
brotlipy=0.7.0=py36h8f6f2f9_1001
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h7f98852_1
ca-certificates=2021.5.30=ha878542_0
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pyhd8ed1ab_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2021.5.30=py36h5fab9bb_0
cffi=1.14.6=py36hc120d54_0
chardet=4.0.0=py36h5fab9bb_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=py36h5fab9bb_0
cryptography=3.4.7=py36hb60f036_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
dataclasses=0.8=pyh787bdff_0
decorator=4.4.2=py_0
dgl=0.6.1=pypi_0
dgl-cuda10.1=0.6.1=py36_0
easydict=1.9=pypi_0
ffmpeg=4.3.1=hca11adc_2
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
gast=0.2.2=pypi_0
gettext=0.19.8.1=h0b5b191_1005
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.32.1=pyh6c4a22f_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pypi_0
googledrivedownloader=0.4=pypi_0
greenlet=1.1.0=py36hc4f0c31_0
grpcio=1.38.1=py36h8e87921_0
h5py=3.1.0=pypi_0
icu=68.1=h58526e2_0
idna=3.1=pyhd3deb0d_0
idna_ssl=1.1.0=py36h9f0ad1d_1001
importlib-metadata=4.6.1=py36h5fab9bb_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
isodate=0.6.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py36h605e78d_1
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.36.1=hea4e1c9_1
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_19
libgfortran-ng=7.5.0=h14aa051_19
libgfortran4=7.5.0=h14aa051_19
libglib=2.68.3=h3e27bee_0
libiconv=1.16=h516909a_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.17.2=h780b84a_1
libstdcxx-ng=9.3.0=h6de172a_19
libtiff=4.2.0=h85742a9_0
libuuid=2.32.1=h7f98852_1000
libuv=1.41.1=h7f98852_0
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxml2=2.9.12=h72842e0_0
littleutils=0.2.2=pypi_0
llvm-openmp=12.0.1=h4bd325d_0
lz4-c=1.9.3=h9c3ff4c_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=2.0.1=pypi_0
matplotlib-base=3.3.4=py36hd391965_0
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h92226af_1
mkl_random=1.1.1=py36h0573a6f_0
mpmath=1.2.1=pypi_0
multidict=5.1.0=py36h8f6f2f9_1
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
networkx=2.5.1=pypi_0
ninja=1.10.2=h4bd325d_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.1=pyhd8ed1ab_0
ogb=1.3.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1k=h7f98852_0
opt-einsum=3.3.0=pypi_0
outdated=0.2.1=pypi_0
pandas=1.1.5=pypi_0
parso=0.8.2=pypi_0
pbr=5.6.0=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py36h2c7a002_0
pip=21.1.3=pypi_0
pixman=0.40.0=h36c2ea0_0
prompt-toolkit=3.0.19=pypi_0
protobuf=3.17.2=py36hc4f0c31_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
py4j=0.10.9=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycairo=1.20.1=py36h726f1de_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.9.0=pypi_0
pyjwt=2.1.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysmiles=1.0.1=pypi_0
pysocks=1.7.1=py36h5fab9bb_3
pyspark=3.1.2=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.15=pypi_0
python_abi=3.6=2_cp36m
pytorch=1.8.1=py3.6_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
rdflib=5.0.0=pypi_0
rdkit=2021.03.3=py36h59cdaa9_0
readline=8.1=h46c0cb4_0
reportlab=3.5.68=py36h3e18861_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pypi_0
setuptools=49.6.0=py36h5fab9bb_3
six=1.16.0=pyh6c4a22f_0
sklearn=0.0=pypi_0
sqlalchemy=1.4.21=py36h8f6f2f9_0
sqlite=3.36.0=h9cd32fc_0
sympy=1.8=pypi_0
tensorboard=1.15.0=pypi_0
tensorboard-data-server=0.6.0=py36h0a59100_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorboardx=2.0=pypi_0
tensorflow=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=h21135ba_1
torch-cluster=1.5.9=pypi_0
torch-geometric=1.7.2=pypi_0
torch-scatter=2.0.7=pypi_0
torch-sparse=0.6.10=pypi_0
torch-spline-conv=1.2.1=pypi_0
torch-tb-profiler=0.2.0=pypi_0
torchaudio=0.8.1=py36
torchvision=0.9.1=py36_cu102
tornado=6.1=py36h8f6f2f9_1
tqdm=4.61.1=pypi_0
traitlets=4.3.3=pypi_0
typing-extensions=3.10.0.0=hd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
urllib3=1.26.6=pyhd8ed1ab_0
warmup-scheduler=0.3=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
wrapt=1.12.1=pypi_0
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yarl=1.6.3=py36h8f6f2f9_2
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstd=1.4.9=ha95c52a_0
